+++ src/mkvtoolnix-gui/main_window/preferences_dialog.h
@@ -77,6 +77,7 @@ protected:
 
   void setupInterfaceLanguage();
   void setupTabPositions();
+  void setupDerivingTrackLanguagesFromFileName();
   void setupWhenToSetDefaultLanguage();
   void setupJobRemovalPolicy();
   void setupCommonLanguages();